streamlit/requirements.txt
streamlit>=1.35.0
psycopg2-binary>=2.9.9
pandas>=2.2.0
plotly>=5.22.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
